+++ instrumentation/opentelemetry-instrumentation-requests/src/opentelemetry/instrumentation/requests/__init__.py
@@ -36,18 +36,17 @@ API
 import functools
 import types
 
-from requests import Timeout, URLRequired
-from requests.exceptions import InvalidSchema, InvalidURL, MissingSchema
 from requests.models import Response
 from requests.sessions import Session
 from requests.structures import CaseInsensitiveDict
 
-from opentelemetry import context, propagators
+from opentelemetry import context
 from opentelemetry.instrumentation.instrumentor import BaseInstrumentor
 from opentelemetry.instrumentation.requests.version import __version__
 from opentelemetry.instrumentation.utils import http_status_to_status_code
+from opentelemetry.propagate import inject
 from opentelemetry.trace import SpanKind, get_tracer
-from opentelemetry.trace.status import Status, StatusCode
+from opentelemetry.trace.status import Status
 
 # A key to a context variable to avoid creating duplicate spans when instrumenting
 # both, Session.request and Session.send, since Session.request calls into Session.send
@@ -135,7 +134,7 @@ def _instrument(tracer_provider=None, span_callback=None, name_callback=None):
                 span.set_attribute("http.url", url)
 
             headers = get_or_create_headers()
-            propagators.inject(type(headers).__setitem__, headers)
+            inject(type(headers).__setitem__, headers)
 
             token = context.attach(
                 context.set_value(_SUPPRESS_REQUESTS_INSTRUMENTATION_KEY, True)