+++ sdk-extension/opentelemetry-sdk-extension-aws/src/opentelemetry/sdk/extension/aws/trace/propagation/aws_xray_format.py
@@ -38,10 +38,10 @@ Or by setting this propagator in your instrumented application:
 
 .. code-block:: python
 
-    from opentelemetry import propagators
+    from opentelemetry.propagate import set_global_textmap
     from opentelemetry.sdk.extension.aws.trace.propagation.aws_xray_format import AwsXRayFormat
 
-    propagators.set_global_textmap(AwsXRayFormat())
+    set_global_textmap(AwsXRayFormat())
 
 API
 ---